Add support for repos out of litex-hub.
Not all IP is ready for general use, so allowing this
tool to target things outside the litex-hub repo is
a decent halfway point for things not yet ready (or
will never graduate past toy CPU).

We probably want to support having a maintainer besides
'LiteX Authors' for these too.